Шикра, Комета, 30мм корабельный пулемёт

+++ code/modules/overmap/sectors.dm
@@ -38,8 +38,13 @@
 
 	forceMove(locate(start_x, start_y, GLOB.using_map.overmap_z))
 
+	var/obj/effect/overmap/visitable/S
+
 	for(var/obj/effect/overmap/event/E in loc)
-		qdel(E)
+		if(E == /obj/effect/overmap/event/star)
+			S.forceMove(locate(rand(OVERMAP_EDGE, GLOB.using_map.overmap_size - OVERMAP_EDGE), rand(OVERMAP_EDGE, GLOB.using_map.overmap_size - OVERMAP_EDGE), GLOB.using_map.overmap_z))
+		else
+			qdel(E)
 
 	docking_codes = "[ascii2text(rand(65,90))][ascii2text(rand(65,90))][ascii2text(rand(65,90))][ascii2text(rand(65,90))]"
